+++ changelog.md
@@ -10,6 +10,7 @@ Given a version number MAJOR.MINOR.PATCH, increment the:
 Additional labels for pre-release and build metadata are available as extensions to the MAJOR.MINOR.PATCH format. In our case this is e.g. `-rc.1`, `-rc.2`.
 
 ## Unreleased changes
+* [Issue 70](https://github.com/AmpersandTarski/prototype/issues/70) Improve exec engine error message
 * UI Fix for longer notification (error) messages. Message now overflows on new line
 * Add missing parameter and return type declarations for all methods. Remove unneeded docblock comments
 * Add class property type declarations. Remove unneeded docblock comments